Connect account picture menu action
